feat(tools): 新增「离线密码重置」工具，并将 SAM 改密逻辑提取到 lr-core 共享
- 新增共享模块 lr-core::sam：离线清除指定账户密码 + 启用被禁用账户
  （clear_account_password 及全部 V/F 解析辅助），含强制备份 SAM、成功后删除
  备份等安全措施；7 个纯逻辑单元测试一并随测试搬入（随 cargo test -p lr-core 跑）。
- PE端 account_fix：删除重复的 SAM 实现与测试，改调 lr_core::sam::clear_account_password
  （ensure_offline_login 行为不变），去重约 200 行。
- 正常系统端工具箱新增「离线密码重置」：输入离线 Windows 盘符 + 用户名，后台调用
  lr_core::sam 清除该账户密码（可空密码登录）并启用账户；新增 ui/tools/password_reset.rs、
  app 状态、按钮与状态轮询接线，带「需授权场景/会修改 SAM」警示。

至此本会话承诺的新功能全部完成：SHA-256 校验、BitLocker 恢复密钥备份、
BitLocker 挂起/恢复、离线密码重置；以及 wimlib 满核+高优先级性能改动。
